# openapi.yaml
openapi: 3.0.0
info:
  title: saludconecta API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3015/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /cita-medica:
    get:
      tags:
        - CitaMedica
      summary: Listar todos los CitaMedicas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de CitaMedicas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - CitaMedica
      summary: Crear CitaMedica
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: CitaMedica creado
  /cita-medica/{id}:
    get:
      tags:
        - CitaMedica
      summary: Obtener CitaMedica por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CitaMedica encontrado
        '404':
          description: CitaMedica no encontrado
    patch:
      tags:
        - CitaMedica
      summary: Actualizar CitaMedica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: CitaMedica actualizado
    delete:
      tags:
        - CitaMedica
      summary: Eliminar CitaMedica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: CitaMedica eliminado
  /cita-medica/{id}/upload-image:
    post:
      tags:
        - CitaMedica
      summary: Subir imagen a CitaMedica
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /historial-clinico:
    get:
      tags:
        - HistorialClinico
      summary: Listar todos los HistorialClinicos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de HistorialClinicos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - HistorialClinico
      summary: Crear HistorialClinico
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: HistorialClinico creado
  /historial-clinico/{id}:
    get:
      tags:
        - HistorialClinico
      summary: Obtener HistorialClinico por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HistorialClinico encontrado
        '404':
          description: HistorialClinico no encontrado
    patch:
      tags:
        - HistorialClinico
      summary: Actualizar HistorialClinico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: HistorialClinico actualizado
    delete:
      tags:
        - HistorialClinico
      summary: Eliminar HistorialClinico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: HistorialClinico eliminado
  /historial-clinico/{id}/upload-image:
    post:
      tags:
        - HistorialClinico
      summary: Subir imagen a HistorialClinico
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...